+++ packages/ui-react/src/components/modal/Modal.tsx
@@ -24,6 +24,9 @@ function Modal(props: ModalProps, ref?: ForwardedRef<HTMLDivElement>) {
     center = false,
     extra,
     open,
+    widthClamp,
+    omitContentStyle,
+    clampHeightByWindow,
     ...rest
   } = props;
   const { onClose, onBack } = props;
@@ -54,12 +57,25 @@ function Modal(props: ModalProps, ref?: ForwardedRef<HTMLDivElement>) {
   );
 
   return (
-    <ModalOverlay in={open} {...rest} ref={ref}>
-      <ModalStyle $center={center}>
-        <ModalBaseStyle>
+    <ModalOverlay
+      in={open}
+      {...rest}
+      ref={ref}
+      clampHeightByWindow={clampHeightByWindow}
+    >
+      <ModalStyle
+        $center={center}
+        $width={widthClamp}
+        $clampHeightByWindow={clampHeightByWindow}
+      >
+        <ModalBaseStyle $clampHeightByWindow={clampHeightByWindow}>
           {modalHeader}
           {withSubtitle && <ModalSubtitleStyle>{subtitle}</ModalSubtitleStyle>}
-          <ModalContentStyle>{children}</ModalContentStyle>
+          {omitContentStyle ? (
+            children
+          ) : (
+            <ModalContentStyle>{children}</ModalContentStyle>
+          )}
         </ModalBaseStyle>
         {extra}
       </ModalStyle>