+++ packages/ui-react/src/components/modal/types.tsx
@@ -16,7 +16,9 @@ export type ModalOverlayOwnProps = CommonComponentProps<
 
 export type ModalOverlayProps = ModalOverlayOwnProps & TransitionWrapperProps;
 export type ModalOverlayInnerProps = ModalOverlayOwnProps &
-  TransitionInnerProps;
+  TransitionInnerProps & {
+    clampHeightByWindow?: boolean;
+  };
 
 export type ModalProps = {
   title?: React.ReactNode;
@@ -25,6 +27,9 @@ export type ModalProps = {
   extra?: React.ReactNode;
   center?: boolean;
   open?: boolean;
+  omitContentStyle?: boolean;
+  widthClamp?: number;
+  clampHeightByWindow?: boolean;
 } & Omit<ModalOverlayProps, 'title' | 'in'>;
 
 export type ModalExtraProps = CommonComponentProps<'div'>;